moved somethings into util package and made day one fit naming convention for previous days
