+++ 2019/1/lib/lib_test.go
@@ -1,4 +1,4 @@
-package one
+package lib
 
 import "testing"
